Split neighborhood_quality.rs (844 lines) into subdirectory module
Break the monolithic neighborhood_quality.rs into focused sub-modules:
- types.rs (110 lines): DistrictQuality, NeighborhoodQualityIndex, weight constants, Saveable impl
- compute.rs (217 lines): pure computation functions (walkability, service coverage, environment, safety, park access, composite index)
- systems.rs (95 lines): ECS update_neighborhood_quality system
- tests.rs (418 lines): all unit tests
- mod.rs (55 lines): module declarations, re-exports, plugin registration

All pub items are re-exported from mod.rs for backward compatibility.
No functional changes.

Part of #1389
